:sparkles: [feat] 프로필 영상 클릭시 자신이 올린 영상과 다른 사용자가 올린 영상에 대해 다른 디자인의 screen 보이는 기능 개발 (HH-323)
